Use docker api stats stream to fetch stats
This is a bit different from the previous approach of getting the list of running containers, the details and the stats on each request to the metrics endpoint. This change modifies that behaviour and introduces two types of threads that constantly updates the stats of the container using a Stream, which is a lot faster to consume that to create a new connection every time.

The first type of Thread (DockerStatsCollector) will maintain updated the lists of containers, when a new container is launched, a new thread of the second type (ContainerStatsStream) is launched to read from the stream. When the metrics are requested, the collector will read the latest entry of each container and parse that.

This approach do introduces load over time, even if the metrics endpoint is not queried, but it saves a lot of time on requests.